Redirect to ban appeal if the user is banned
The default behavior is for the `/session/` request to be redirected to the ban appeal page. Unfortunately we can't detect this 302 response, as the browser transparently handles it and javascript just sees a 200 response with a weird body.  So I've updated scratchr2 to return a special response for banned `/session/` requests.