Expand view to wrap contents on small screens
`inline-block` causes #view width to expand to its children's width.

`min-width: 100%` is necessary for larger screens, so that the `inline-block` element is still grows to contain all the available width.

The 1px of padding is apparently needed for Android Chrome, otherwise the right edge has blue from the body background showing.