Use xhr to clear cache instead of form
because csrf. also, since we don't have a definite message/banner system in place yet, use the button itself for feedback rather than messages. The button will disable if it has been successfully used, or turn pink if there was an error (and it will not be disabled in that case).