Set up Raven within init.js
Now that Raven is bundled with webpack, it's not available in the global context, so require it and configure it within the init module.

I struggled to figure out how to expose the raven-js module as a global, as all the combinations of `externals` and `ProvidePlugin` only applied to webpacked modules, not the global window.  The correct way to do it seems to be exports-loader but that looked rather hacky and hard to understand.