Fix search button display
Update search form submit button from an input[type=submit] to a button.  React 15 omits the value prop if it's blank, so the `<input type="submit" value="">` was rendered as `<input type="submit">`. This caused the input to display the default value of "Submit".  The button element is more appropriate for this behavior and doesn't have the missing value issue.