Add webkit prefixes for safari 8
`caniuse` updated their api recently, and it appears that support for safari 8 is no longer in the last 3 versions (and it appears that safari 8 is not listed as supported there in general). explicitly add the version back in to make sure the proper `-webkit` prefixes are included for that browser.