Add dependency for node-dir 0.1.16 (#1354)
FIxes problem where webpack wasn’t copying static or intl to the build directory.

Gory details:
node-dir is a dependency of copy-webpack-plugin. If you don’t already have it installed, the latest version (0.1.17) gets installed and that one breaks copy-webpack (https://github.com/fshost/node-dir/issues/50). Webpack build will complete. but there will be an error `ERROR in this is not a typed array.` We didn’t see it earlier because node-modules are cached.