Hide Firefox focus ring
See this StackOverflow answer: https://stackoverflow.com/a/18853002/4633828

This isn't a perfect fix - we have to set `color` explicitly, rather
than use `inherit`, because we want the text color to be the same when
the select element is focused. That means both `color` and
`text-shadow`'s color have to be explicitly set to the same thing
(otherwise they might not match). Yay, hacks!