Linting tests will not work if scratch-gui is linked directly into local directory
We’re currently linking scratch-gui directly into the scratch-www directory (either through a bind mount in docker) or a symoblic link in unix. This causes the current make lint target to fail when it attempts to look in that directory for lint targets. Also, don’t bother linting node_modules