Prevent uncaught errors from going to sentry
For real this time. Just needed to figure out the right plugins to filter. TryCatch plugin is the one that reports errors from setTimeouts, which the VM executes code in.

Remove the previous window.onerror kludge, it was not working and prevented errors from reaching the console.

Tested locally by providing the staging DSN for Sentry