Reduce locale data
There is a trade off between adding this data for all users that load the page regardless of what browser they’re using and whether they get muted, and only localizing the most common languages.
In the last 30 days only ~3% of users on scratch use a version of Safari (13.x) that doesn’t support Intl. Of that 3%, these locales would cover over 60% of the users. English will be used by default if the locale data is not available.