add macOS troubleshooting info on all extension pages
I created a new `ExtensionTroubleshooting` component to store shared
troubleshooting steps. In the future I plan to move existing shared
troubleshooting steps into this component, but I don't think it's a good
idea to do this so close to the upcoming Scratch Conference. Moving
those steps will disrupt their localization strings in at least three
ways:
* each page is currently its own "resource" so the strings will move
  from one resource to another
* some of the strings include the device name, which I plan to replace
  using a `formatMessage` variable, changing the string in the process
* reusing the existing string IDs won't make sense: I plan to replace
  `microbit.checkOSVersionTitle` with `extensions.checkOSVersionTitle`
  for example

For now, all existing troubleshooting steps are being treated as
extension-specific in order to avoid localization disruptions. The new
versions of the strings are present, though, in order to allow
translators to start working on them if they'd like.