test: simplify clickXpath() by letting click() scroll
Instead of trying to proactively scroll if necessary, just let Selenium
handle scrolling. The only reason to handle scrolling manually was to
tell the difference between an off-screen element and one that isn't
visible due to being beneath another element (like the loading screen).
Catching `ElementClickInterceptedError` is a simpler way to do that.