ci: remove workflow_run CI/CD trigger, add deploy info
The `workflow_run` trigger ran this workflow at the right time but not
in the right place. It correctly ran after the release branch creation
workflow finished, but it ran in the branch selected in the GitHub UI.
That branch can't be the release branch since it doesn't exist at that
time.