Added rounded corners to "What's happening?" user icons.
There are many places in the website where user icons have a border-radius property of 3 pixels, however the What's happening? section has fully squared borders that do not match the rest. It can be noticed with high zoom.

Padding was replaced by margin because it prevented the border radius from working correctly and was redundant. Replacing it with margin does not affect anything and is in fact cleaner.

Works for all devices and browsers.