fix: fix project page on Safari mobile
This appears to be a general timing issue. I'm not sure why it shows up
as a problem on mobile Safari specifically.

Basically, we need to fetch project info in order for the project page
to load correctly. Previously, that was only done when the session info
changed to "FETCHED" from any other state (or when the project ID
changes). This change also fetches project info when the session info
when the session status is already "FETCHED" before the <Preview>
component is mounted.